backend/requirements.txt
django==2.2
djangorestframework==3.9.3
markdown==3.1
django-filter==2.1.0
coverage==4.5.3
django-cors-headers==2.5.3
djangorestframework-jwt==1.11.0
